Remove the manager concept in favor of client. Remove Symbol.dispose because it doesn't make sense
